AP-12740: Example input data (optionally) stored with component
If component is a project (i.e. not embedded in a workflow). Data is
stored in an additional folder within the component directory.

Example input data info (file names, port types etc.) and flow variables
are stored in template.knime.

Additional tweaks such as
* correct message when no example data is there
* give component input access to that data
* correct write protection set
* ...

AP-12740 (Component editor with example data)